+++ vcert/parser/utils.py
@@ -18,8 +18,7 @@ import logging as log
 import os
 
 from vcert.parser import *
-from vcert.policy import DEFAULT_CA
-from vcert.policy.policy_spec import Subject, KeyPair, SubjectAltNames, Policy, DefaultSubject, DefaultKeyPair, \
+from vcert.policy import DEFAULT_CA, Subject, KeyPair, SubjectAltNames, Policy, DefaultSubject, DefaultKeyPair, \
     Defaults, PolicySpecification
